+++ doc/CHANGELOG.md
@@ -1,3 +1,18 @@
+# v0.3.2 beta
+- New NNX Prompt and add NNX Private
+- Sandbox command (only work for NNX Private)
+- Config.py split into a folder
+- Fix bug when input the wrong command or the command has an error, you will be kicked out of cmd.
+- Fix Plugin command error nofication
+(NNX Plugin Store is cancelled because unstable)
+
+# v0.3.2exp beta (including v0.3.1.1exp and v0.3.1.2exp)
+- New prompt
+- Simplified configuration process for easier setup
+- Fixed some bug and improve stability
+- Not adding new command btw
+(Note: NNX Plugin Store is currently unavailable for download due to some issues, will be fix soon)
+
 # v0.3.1 beta
 - Add process command (two command btw)
 - Good showing error
@@ -9,3 +24,5 @@
 - Add new two commands
 - Fully fixed not showing error (or better for your eyes)
 - Fixed some bugs
+
+# v0.2.10exp beta: Added monitoring system and fix some bug